Time: 168 ms (45.86%), Space: 6.3 MB (48.87%) - LeetHub

+++ 300-longest-increasing-subsequence/300-longest-increasing-subsequence.c
@@ -2,11 +2,11 @@
 #define min(a,b) (a<b?a:b)
 
 int lengthOfLIS(int* nums, int n){
-    int i,j;
+    short i,j;
     int max_lis;
-    int *lis;
+    short *lis;
     
-    lis = malloc(n*sizeof(int));
+    lis = malloc(n*sizeof(short));
     if (lis == NULL) {
         return 0;
     }
